prober: fix 1B probes limit overflow

+++ src/heartbeat_config.hpp
@@ -16,7 +16,7 @@ struct HeartbeatConfig {
   const optional<fs::path> input_file;
   const fs::path output_file;
   const optional<fs::path> start_time_log_file;
-  const optional<int> max_probes;
+  const optional<unsigned long long int> max_probes;
   const int n_packets;
   const int probing_rate;
   const int sniffer_buffer_size;
@@ -68,7 +68,7 @@ class HeartbeatConfigBuilder {
     m_sniffer_buffer_size = size;
   }
 
-  void set_max_probes(const int count) {
+  void set_max_probes(const unsigned long long int count) {
     if (count <= 0) {
       throw std::domain_error("max_probes must be > 0");
     }
@@ -181,7 +181,7 @@ class HeartbeatConfigBuilder {
   optional<string> m_protocol;
   optional<string> m_interface;
   optional<int> m_sniffer_buffer_size;
-  optional<int> m_max_probes;
+  optional<unsigned long long int> m_max_probes;
   optional<int> m_n_packets;
   optional<fs::path> m_bgp_filter_file;
   optional<fs::path> m_prefix_excl_file;
